rec: move contexts belonging to a Rec to rec
Since the current implementations of Timer, Gic, and Simd
serve as parts of Rec's context, move them under rec.
