+++ rmm/src/rsi/mod.rs
@@ -16,10 +16,10 @@ use crate::granule::{GranuleState, GRANULE_SIZE};
 use crate::listen;
 use crate::measurement::{HashContext, Measurement, MEASUREMENTS_SLOT_NR, MEASUREMENTS_SLOT_RIM};
 use crate::realm::config::realm_config;
-use crate::realm::context::{get_reg, set_reg};
 use crate::realm::mm::address::GuestPhysAddr;
 use crate::realm::mm::rtt::RTT_PAGE_LEVEL;
 use crate::realm::rd::Rd;
+use crate::rec::context::{get_reg, set_reg};
 use crate::rec::{Rec, RmmRecAttestState};
 use crate::rmi;
 use crate::rmi::error::Error;
